nodejs 설치에 대한 y/n 추가

+++ script/install_neovim.sh
@@ -17,7 +17,7 @@ make CMAKE_BUILD_TYPE=Release CMAKE_INSTALL_PREFIX=/opt/nvim
 RET_CODE="$?"
 if [ $RET_CODE -ne 0 ]; then
     echo "compile failed... [$RET_CODE]"
-	exit $RET_CODE
+    exit $RET_CODE
 fi
 
 mkdir -p /opt/nvim
@@ -31,23 +31,36 @@ if [ -f "/opt/nvim/bin/nvim" ]; then
     cd ../
     rm -rf neovim
 
-	echo "neovim compile done"
-	nvim --version
+    echo "neovim compile done"
+    nvim --version
 fi
 
 
 if [ -z "$(command -v node)" ]; then
+
+    echo -n "not found nodejs. install nodejs? [y/n]: "
+    read answer
+    if [[ ! "$answer" =~ ^[yY](es)?$ ]]; then
+        echo "nodejs install cancelled."
+        echo "you should install tree-sitter-cli."
+        exit 1
+    fi
+    echo -n "input nodejs version to install (18/20/22/24... default: 22): "
+    read version
+    NODE_MAJOR=${version:-22}
+
     apt-get update
     apt-get install -y ca-certificates curl gnupg
     mkdir -p /etc/apt/keyrings
     curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
-
-    NODE_MAJOR=22
     echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | tee /etc/apt/sources.list.d/nodesource.list
     unset NODE_MAJOR
 
     apt-get update
     apt-get install -y nodejs
+
+    npm install -g tree-sitter-cli
+
 fi
